update httpclient tests
