--- src/Client/RestClientSettings.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace BlazorFocused.Client
-{
-    public class RestClientSettings
-    {
-        public string BaseAddress { get; set; }
-    }
-}
